+++ src/Common/JK.Common/Data/AdoBaseOperation.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Data;
+using System.Data.Common;
+
+namespace JK.Common.Data
+{
+    [Obsolete("Generic ADO classes have not been tested in .NET Core. Left in for reference purposes.")]
+    public abstract class AdoBaseOperation
+    {
+        private readonly AdoBaseRepository repository;
+
+        /// <summary>
+        /// Initializes a new instance of the AdoBaseOperation class.
+        /// </summary>
+        /// <param name="adoBaseRepository">The ADO base repository to use.</param>
+        protected AdoBaseOperation(AdoBaseRepository adoBaseRepository)
+        {
+            this.repository = adoBaseRepository;
+        }
+
+        protected AdoBaseRepository Repository => this.repository;
+
+        protected abstract void SetupParameters(DbCommand command);
+
+        protected AdoParameterFactory GetParameterFactory(IDbCommand command)
+        {
+            return this.repository.GetParameterFactory(command);
+        }
+
+        protected DbCommand SetupCommand(DbConnection connection, CommandType commandType, string commandText)
+        {
+            DbCommand command = connection.CreateCommand();
+            command.CommandType = commandType;
+            command.CommandText = commandText;
+            this.SetupParameters(command);
+            return command;
+        }
+    }
+}